add IDs for each query task
a race condition existed where MSqlQuery::resultsReady() signal gets
emitted twice (when calling execAsync() twice on the same MSqlQuery
object, and blocking the thread between both calls).

this commit adds an ID for each query task, to guarantee that
resultsReady() signal is emitted only for the last set task.

queries that are set while the MSqlQuery is busy, will either be
overwritten (never executed), or executed silently (no signal will be
emitted for them, and there is no way to know whether they have been
executed successfuly or not).